Support x64 tests
This fixes various tests that were not compatible with x64. The real
changes are to tests that create/map/recycle frames and these have
been rewritten to be clearer and support both 32 and 64bit.

Other tests have minor changes for type compatibility and reserving
additional amounts of memory needed for the changes.

+++ apps/sel4test-tests/src/tests/domains.cxx
@@ -75,7 +75,7 @@ int fdom1(seL4_Word id, env_t env)
     while (countdown > 0) {
         wait(POLL_DELAY_NS, env);
         --countdown;
-        ZF_LOGD("%2d, ", id);
+        ZF_LOGD("%2d, ", (int)id);
     }
 
     return sel4test_get_result();
